requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.10
beautifulsoup4==4.13.3
pytest==7.1.3
coverage==7.8.0